job_search_system/requirements.txt
# Job Search System Dependencies
# All free and open-source libraries

# Web scraping and HTTP requests
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data handling
pandas==2.2.0
openpyxl==3.1.2

# Email handling
dnspython==2.6.1
email-validator==2.1.0

# Scheduling
schedule==1.2.1
python-crontab==3.0.0

# Logging
colorlog==6.8.2

# Environment variables
python-dotenv==1.0.1

# Rate limiting
ratelimit==2.2.1

# Text processing
jinja2==3.1.3

# HTML parsing
html5lib==1.1
